requirements.txt
# Core data science libraries
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.6
joblib==1.1.0

# Visualization
matplotlib==3.5.1
seaborn==0.11.2
plotly==5.6.0

# Utilities
pyyaml==6.0
tqdm==4.62.3

# Jupyter
jupyter==1.0.0
ipykernel==6.5.1

# API (optional)
fastapi==0.68.0
uvicorn==0.15.0

# Documentation
sphinx==4.3.1
sphinx-rtd-theme==1.0.0